feat(Lucene Builder): create separate files
